+++ src/server/services/agentSignal/policies/completionPolicy.ts
@@ -1,31 +1,26 @@
 import { AGENT_SIGNAL_SOURCE_TYPES } from '@lobechat/agent-signal/source';
-import type { BuiltinAgentSlug } from '@lobechat/builtin-agents';
-import { SELF_ITERATION_AGENT_SLUGS } from '@lobechat/builtin-agents';
 
 import { defineAgentSignalHandlers, defineSourceHandler } from '../runtime/middleware';
 import type { SelfIterationCompletionPayload } from '../services/selfIteration/completion';
 
 /**
- * Handles `agent.execution.completed` source events emitted after every execAgent run
- * (including builtin background agents). Routes builtin self-iteration runs
- * (nightly-review / self-reflection / self-feedback-intent) to an optional
- * caller callback so side-effects (brief writing, receipt projection,
- * idempotency marker) can happen asynchronously after the agent run finishes.
+ * Handles `agent.execution.completed` source events emitted after every execAgent
+ * run. Routes runs that stamped an agent-signal marker (nightly-review /
+ * self-reflection / self-feedback-intent / memory) to an optional caller
+ * callback so side-effects (receipt projection) can happen asynchronously after
+ * the agent run finishes.
  *
- * Mode is carried by `agentId` itself — each self-iteration mode is a distinct
- * builtin agent slug, so callers can dispatch on `agentId` without needing to
- * read the operation row.
+ * Routing is keyed on the marker-derived `selfIteration` payload, NOT the agent
+ * slug: a memory-writer run executes as the user's own agent, so a slug check
+ * would miss it. The marker is the authoritative "this run wants completion-side
+ * projection" signal, stamped by the dispatcher.
  *
  * The callback is fire-and-forget from the worker's perspective; failures are
  * logged but never re-trigger the source pipeline.
- *
- * NOTE on userId: the `agent.execution.completed` source payload does not carry
- * userId, and `AgentSignalSource.scope` is not populated by renderers. Callers
- * that need userId should look it up via the operations table by `operationId`.
  */
 export interface CompletionCallbackParams {
-  /** Self-iteration agent slug — caller dispatches mode-specific behaviour from this. */
-  agentId: BuiltinAgentSlug;
+  /** Agent that ran — a builtin self-iteration slug, or a user agent for memory. */
+  agentId: string;
   operationId: string;
   /**
    * Compact self-iteration tool outcomes + run marker, extracted from the run's
@@ -54,15 +49,15 @@ export const createCompletionPolicy = (options: CreateCompletionPolicyOptions = 
         const { agentId, operationId, selfIteration, topicId } = source.payload;
 
         if (!agentId || !operationId) return;
-        if (!SELF_ITERATION_AGENT_SLUGS.has(agentId as BuiltinAgentSlug)) return;
+        // Marker-driven: only runs that stamped a marker carry a selfIteration
+        // payload. Unmarked runs have nothing to project.
+        if (!selfIteration) return;
         if (!options.onSelfIterationCompleted) return;
 
         const params: CompletionCallbackParams = {
-          agentId: agentId as BuiltinAgentSlug,
+          agentId,
           operationId,
-          ...(selfIteration
-            ? { selfIteration: selfIteration as SelfIterationCompletionPayload }
-            : {}),
+          selfIteration: selfIteration as SelfIterationCompletionPayload,
           ...(topicId ? { topicId } : {}),
         };
